pluto add sa_priority to raw_eroute to raw_eroute to set priority for in, out and fwd xfrm pol. consistant type for priority and sa_priority. same as in xfrm, uint32_t

+++ programs/pluto/connections.h
@@ -202,7 +202,7 @@ struct connection {
 	deltatime_t sa_rekey_margin;
 	unsigned long sa_rekey_fuzz;
 	unsigned long sa_keying_tries;
-	unsigned long sa_priority;
+	uint32_t sa_priority;
 	unsigned long r_interval; /* initial retransmit time in msec, doubles each time */
 	deltatime_t r_timeout; /* max time (in secs) for one packet exchange attempt */
 	reqid_t sa_reqid;
